pyproject.toml
[project]
name = "webdev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.4",
    "python-fasthtml>=0.10.1",
]
